refactor(compiler): rename ConstantValue
- I've renamed ConstantValue and all the related types to ConstantExpression to better match the
  parser grammar.

Issues: #8
